Add logic for manually specifying the initial crop rect. (#18)
This allows developers to specify the cropping rect rather than using the bounds of the `<img>` element. So now you can do something like:

```html
<div style="overflow: hidden">
  <img src="..." style="position: absolute; top: 0; left: 0; width: 100%; height: 100%; object-fit: cover; transform: scale(2);">
</div>
```

In this example, you will have something that renders like:
```
________________
|     ____     |
|    |    |    |
|    |    |    |                    
|     ‾‾‾‾     |
‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾
```
where the inner rectangle represents the cropping div, and the outer rectangle represents the `<img>` boundaries. With this, you can zoom in on a portion of an image in a way that `object-position` does not allow you to. As an alternative, you could use `object-fit: none`, `object-position`, and `transform: scale`, but this does not work well when trying to responsively size images (e.g. using the `padding-bottom` trick).

Without this change, the animation is off as the image portion escaping the overflow container is visible at the start/end on the first frame of the animation (instead of the crop growing),